fix(lens): share one case-insensitive header lookup (getHeaderCI)
Add `getHeaderCI(headers, name)` in @kolu/padi beside contentLengthOf (derived from it now), re-export it through @kolu/padi/assembly, and have kolu-server's assembleRemotePreview consume it — deleting the hand-rolled `headerValue` and routing totalFromContentRange through the shared helper.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-2, raised by hickey). Not pushed or merged.
